SLVS-2520 Move SelectedItem inside the ReportViewModel as there should be one selection possible in the whole tree (as opposed to one per group)
